service

+++ sanghwa/sanghwa/src/main/java/com/example/sanghwa/repository/LectureRepository.java
@@ -3,6 +3,5 @@ package com.example.sanghwa.repository;
 import com.example.sanghwa.domain.Lecture;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LectureRepository extends JpaRepository<Lecture, Integer> {
-
+public interface LectureRepository extends JpaRepository<Lecture, Long> {
 }
