+++ sanghwa/sanghwa/src/main/java/com/example/sanghwa/repository/StrudentRepository.java
@@ -3,5 +3,5 @@ package com.example.sanghwa.repository;
 import com.example.sanghwa.domain.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StrudentRepository extends JpaRepository<Student, Integer> {
+public interface StrudentRepository extends JpaRepository<Student, Long> {
 }
